JavaScript
This part of the website is using JS, but the issue is that in my version of Visual Studio 2017, I'm having trouble getting the js to function on the website. Everything is in place to work, I just may have made an error somewhere. Overall I really enjoyed learning from this project and hope to make much more robust websites in the future!.